refactor: 주별, 3일보기 시 반복되는 영역 component 분리
