+++ client/components/calendar/WeekWrap.tsx
@@ -1,28 +1,19 @@
-import {useRouter} from 'next/router';
-
 import styled from 'styled-components';
 
 import {
-    useRecoilState,
     useRecoilValue,
-    useSetRecoilState
 } from 'recoil';
+
 import {
     targetStateState,
-    todayState,
     viewState
 } from '../../recoil/atoms';
 
 import {
-    isTodayValue,
-    SetDateType,
     ViewType
 } from '../../utils/constants';
 
-import {useChangeDay} from '../../hooks/useChangeDate';
-
-import {Num} from './Num';
-import {TimelineComponent} from './Timeline';
+import {WeekComponent} from './Week';
 
 interface WeekType {
     type: string
@@ -31,19 +22,12 @@ interface WeekType {
 export const WeekWrapComponent = ({
     type
 }: WeekType) => {
-    const router = useRouter();
-
-    const today = useRecoilValue(todayState);
-
-    const [curr, setCurr] = useRecoilState(targetStateState);
-    const [view, setView] = useRecoilState(viewState);
+    const curr = useRecoilValue(targetStateState);
+    const view = useRecoilValue(viewState);
 
     const {
-        fullYear,
         month,
-        date,
         weekFirstNumber,
-        monthLastNumber,
         monthPrevLastNumber,
     } = curr;
 
@@ -61,62 +45,11 @@ export const WeekWrapComponent = ({
         return new Array(nextCount).fill(1).reduce((acc, curr, i) => [...acc, curr + i], []);
     };
 
-    return (<>
-            {view.type !== ViewType.Day && <StyledWeeks>
-                {view.type === ViewType.Week && arrayPrev().map((w: number) => <StyledWeek key={`week_${w}`}>
-                    <StyledNumWrap>
-                        <Num onClick={() => {
-                            useChangeDay({
-                                currMonth: month - 1,
-                                currYear : fullYear,
-                                currDate: w,
-                                setCurr,
-                                setView,
-                                router
-                            });
-                        }}
-                             isToday={isTodayValue(today, fullYear, month - 1, +w)}>{w}</Num>
-                    </StyledNumWrap>
-                </StyledWeek>)}
-
-                {arrayCurrent().map((w: number, index: number) => <StyledWeek key={`week_${w}`}>
-                    <StyledNumWrap>
-                        <Num onClick={() => {
-                            useChangeDay({
-                                currMonth: month,
-                                currYear : fullYear,
-                                currDate: w,
-                                setCurr,
-                                setView,
-                                router
-                            });
-                        }}
-                             isToday={isTodayValue(today, fullYear, month, +w)}>{w}</Num>
-                    </StyledNumWrap>
-                    <TimelineComponent fullYear={fullYear}
-                                       month={month}
-                                       date={+w}
-                                       isToday={isTodayValue(today, fullYear, month, +w)}/>
-                </StyledWeek>)}
-
-                {arrayNext().map((w: number, index: number) => <StyledWeek key={`week_${w}`}>
-                    <StyledNumWrap>
-                        <Num onClick={() => {
-                            useChangeDay({
-                                currMonth: month + 1,
-                                currYear : fullYear,
-                                currDate: w,
-                                setCurr,
-                                setView,
-                                router
-                            });
-                        }}
-                             isToday={isTodayValue(today, fullYear, month + 1, +w)}>{w}</Num>
-                    </StyledNumWrap>
-                </StyledWeek>)}
-
-            </StyledWeeks>}
-        </>
+    return (<StyledWeeks>
+            {view.type === ViewType.Week && <WeekComponent weekDates={arrayPrev()} currMonth={month -1} />}
+            <WeekComponent weekDates={arrayCurrent()} currMonth={month} />
+            <WeekComponent weekDates={arrayNext()} currMonth={month + 1} />
+        </StyledWeeks>
     );
 };
 
@@ -126,53 +59,4 @@ const StyledWeeks = styled.ul`
   display: grid;
   grid-row: 2 / 3;
   align-items: stretch;
-`;
-
-const StyledWeek = styled.li`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  position: relative;
-  text-align: center;
-
-  &:after {
-    content: "";
-    position: absolute;
-    right: 0;
-    top: 0;
-    width: 1px;
-    height: 100%;
-    background-color: var(--light-gray-color);
-  }
-
-  &:nth-child(7) {
-    &:after {
-      display: none;
-    }
-  }
-
-  button {
-    font-size: var(--default-font);
-  }
-`;
-
-const StyledNumWrap = styled.span`
-  display: flex;
-  justify-content: center;
-  position: sticky;
-  top: 35px;
-  width: 100%;
-  background-color: var(--white-color-80);
-  z-index: 1;
-
-  &:after {
-    content: "";
-    position: absolute;
-    top: 100%;
-    left: 0;
-    width: 100%;
-    height: 50px;
-    background: linear-gradient(0deg, rgba(255, 255, 255, 0) 0%, rgba(255, 255, 255, .8) 100%);
-    pointer-events: none;
-  }
 `;
